Create list example

+++ app/src/main/java/ru/falseteam/tasks/realm/model/Task.kt
@@ -4,7 +4,7 @@ import io.realm.RealmObject
 import io.realm.annotations.PrimaryKey
 
 open class Task(
-        @PrimaryKey open var id: Long = 0,
+        @PrimaryKey open var id: Int = 0,
         open var title: String? = null,
         open var text: String? = null,
         open var isComplete: Boolean = false
